requirements.txt
qrcode[pil]>=7.3
Pillow>=10.0
